+++ test/controllers/admin/items_controller_test.rb
@@ -34,7 +34,7 @@ class Admin::ItemsControllerTest < ActionDispatch::IntegrationTest
     delete admin_item_url(@item)
 
     assert_redirected_to item_path @item
-    assert_equal 'This item is frozen and cannot be edited.', flash[:alert]
+    assert_equal I18n.t('items.edit.read_only'), flash[:alert]
 
     @item.read_only = false
     @item.save!