+++ test/models/draft_item_test.rb
@@ -212,13 +212,6 @@ class DraftItemTest < ActiveSupport::TestCase
 
     @collection.read_only = true
     @collection.save!
-    Collection.create!(title: 'Risque fantasy Books',
-                       owner_id: user.id,
-                       restricted: true,
-                       community_id: @community.id)
-    draft_item.assign_attributes(
-      member_of_paths: { community_id: [@community.id], collection_id: [@collection.id] }
-    )
 
     assert_not draft_item.valid?
     assert_equal ['Collection is frozen and not available for deposit'], draft_item.errors.messages[:member_of_paths]